feat: Update setrole command to allow optional role color
The setrole command now allows users to provide an optional role color or hex color when setting or updating a role. If neither a role color nor a hex color is provided, the user will be prompted to provide one. This enhancement provides more flexibility and customization options for managing roles in the Discord server.