feat: Update LoveCalc command to include special case for "sirro" and "molokov_cocktail"
This commit modifies the LoveCalc command to include a special case where if the usernames of both person1 and person2 are either "sirro" and "molokov_cocktail" or "molokov_cocktail" and "sirro", the love percentage is set to 1000 and the loveemote is set to ":heart_eyes:". This change adds a unique behavior for these specific usernames in the LoveCalc command.